+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `clusterlocktimeout` for flocks on parallel actions
 - `searchdomain` to append a domain suffix to the Proxmox hostname when LINSTOR nodes are registered with FQDNs
 - Support for APIVER14
+- support for LINSTOR auth tokens
 
 ## 8.2.0 - 2026-02-24
 ### Added